Add basic structures for game and menus
This adds the basic structure of the game, where the application state
is split from the 'driver' which renders the game and handles user
input. This should allow adding additional renderers (such as OpenGL) in
the future.

This focuses on getting something drawing, so some simple machinery for
menus was included. This means that right now the game will render and
update for the main menu.