Use better cached 'random' values in chunk
This adds a Random field to the `tile.State` structure which holds a
hash of the effective coordinate of the tile. This hash is generated
with the chunk, and is intended to be calculated when loading chunks
instead of being saved alongside it.

The hash is provided by a new package `simplehash` which implements an
immutable hash type, which allows for re-using an intermediate value in
a simpler way.

Some fixes to rendering were made; previously it would use the same tile
for the entire layer, which was hidden by the dynamic nature of the
Random displayer type.

Finally, on selecting "New Game" the coordinates for the 'player' are
initialized to be centered in the chunk on Z layer 33 (the grass layer).