Add grid of 'active' chunks
This adds a grid of active chunks, in preparation for the 'infinite'
world. The renderer currently just renders the current chunk, and no new
chunks are generated as the player moves.

Some refactoring was done to get this to work; Player coordinates now
have a struct which includes the chunk coordinate, and the Update
function handles player movement in a separate function which takes
delta values for the coordinate.