Change world generation to use perlin noise
The world generation now will use perlin noise to generate chunks. A new
flag has been added to set the seed; if not set, a fixed value will be
used. The terminal driver was also updated to properly display liquids,
since the new chunks use liquids.

Doing this revealed a bug in the tile.State Describe function, so that
got a fix as well.