Add caves to world generation
This adds caves to the world generation via a second noise function. To
vary the shape of the caves on each Z level, the coordinates given to
the noise function are scaled, translated, and rotated differently for
each of the Z levels. This gives us a variety of cave looks without
needing to generate a brand new perlin noise generator for each level.

Additionally, because the chunk generator now logs on construction, the
app creation is deferred until the driver is initialized, such that
debug logging is properly captured.